requirements/base.txt
Django==1.8.4
psycopg2==2.6
gunicorn==19.3.0
celery==3.1.18
raven==5.0.0
whitenoise==1.0.3
factory_boy==2.4.1
mock==1.0.1
django-braces==1.4.0
django-suit==0.2.13
django-dbbackup==1.9.0
boto==2.32.1
pillow==2.9.0
requests==2.7.0
djangorestframework==3.2.0
twilio
django-filter==0.11.0
xlwt==0.7.5
django-import-export
